+++ test/HspecMain.hs
@@ -12,6 +12,7 @@ import qualified Spec.UnicodeHandling
 import qualified Spec.DefinitionMetadata
 import qualified Spec.DerivedOutput
 import qualified Spec.GoldenWire
+import qualified Spec.HandlerExceptions
 import qualified Spec.ModernEra
 import qualified Spec.Progress
 import qualified Spec.ProtocolVersionNegotiation
@@ -32,6 +33,7 @@ main = hspec $ do
     Spec.DefinitionMetadata.spec
     Spec.DerivedOutput.spec
     Spec.GoldenWire.spec
+    Spec.HandlerExceptions.spec
     Spec.ModernEra.spec
     Spec.Progress.spec
     Spec.ProtocolVersionNegotiation.spec